requirements.txt
absl-py==2.3.1
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.5
cachetools==5.5.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distlib==0.4.0
dreamsim==0.2.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
facexlib==0.3.0
filelock==3.16.1
filterpy==1.4.5
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.2.3
future==1.0.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.1.7
huggingface-hub==0.34.4
icecream==2.1.5
identify==2.6.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.41.1
lmdb==1.7.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
nodeenv==1.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
open_clip_torch==2.32.0
openai-clip==1.0.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.0.3
peft==0.13.2
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.5.0
propcache==0.2.0
protobuf==5.29.5
psutil==7.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
ruff==0.12.8
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==1.0.19
tokenizers==0.15.2
tomli==2.2.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.37.2
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
virtualenv==20.33.1
wcwidth==0.2.13
Werkzeug==3.0.6
xxhash==3.5.0
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
